+++ linen-modern/src/main/scala/x7c1/linen/modern/accessor/unread/RawSourceAccessor.scala
@@ -1,5 +1,6 @@
-package x7c1.linen.modern.accessor
+package x7c1.linen.modern.accessor.unread
 
+import x7c1.linen.modern.accessor.LinenOpenHelper
 import x7c1.linen.modern.init.dev.SourceTitle
 import x7c1.wheat.macros.logger.Log
