+++ linen-modern/src/main/scala/x7c1/linen/modern/display/unread/OutlineRowAdapter.scala
@@ -3,8 +3,7 @@ package x7c1.linen.modern.display.unread
 import android.support.v7.widget.RecyclerView.Adapter
 import android.view.ViewGroup
 import x7c1.linen.glue.res.layout.{UnreadOutlineRow, UnreadOutlineRowEntry, UnreadOutlineRowFooter, UnreadOutlineRowSource}
-import x7c1.linen.modern.accessor.unread.{EntryContent, SourceHeadlineContent}
-import x7c1.linen.modern.accessor.{EntryAccessor, SourceKind}
+import x7c1.linen.modern.accessor.unread.{EntryAccessor, SourceKind, EntryContent, SourceHeadlineContent}
 import x7c1.linen.modern.struct.UnreadOutline
 import x7c1.wheat.ancient.resource.ViewHolderProvider
 import x7c1.wheat.macros.logger.Log
